stash for classes move
